Refactor into source code and tests.rs
