+++ src/table/mod.rs
@@ -1,0 +1,564 @@
+use std::cmp::{min, Ordering};
+use std::fmt;
+use std::fmt::Formatter;
+use std::slice::Iter;
+use std::sync::Arc;
+
+use json::{JsonValue, object};
+use poker::{Card, Evaluator};
+
+use crate::actions::HandAction;
+use crate::bet_stage::BetStage::{Flop, PreFlop, River};
+use crate::bet_stage::BetStage;
+use crate::player_components::{ActiveState, Player, PlayerState};
+
+#[cfg(test)]
+mod tests;
+
+pub struct Table {
+    /// All players
+    players: Vec<Player>,
+    /// The hand evaluator
+    evaluator: Arc<Evaluator>,
+    /// The flop cards on the table (None if not dealt yet)
+    flop: Option<[Card; 3]>,
+    /// The turn card on the table (None if not dealt yet)
+    turn: Option<Card>,
+    /// The river card on the table (None if not dealt yet)
+    river: Option<Card>,
+    /// Where the current dealer button is, informs turn order
+    dealer_button_index: usize,
+    /// The size of the ante
+    ante: i32,
+    /// How many hands have been played so far
+    hand_number: i32,
+    /// Whose turn it is right now
+    current_player_index: usize,
+    /// State needed for table betting information
+    table_state: BetStage,
+    /// Player bets, how much each player has bet so far
+    player_bets: Vec<i32>,
+    /// How frequently (after "ante_round_increase" rounds) the ante should be increased
+    ante_round_increase: i32,
+}
+
+impl fmt::Display for Table {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        let player_strings: Vec<_> = self.players.iter().map(|x| x.as_json()).collect();
+        let json_object = object! {
+            flop: self.get_flop_string(),
+            turn: self.get_turn_string(),
+            river: self.get_river_string(),
+            dealer_button_index: self.dealer_button_index,
+            players: player_strings,
+            hand_number: self.hand_number,
+        };
+        write!(f, "{}", json_object.dump())
+    }
+}
+
+impl Table {
+    const ANTE_INCREASE_AMOUNT: i32 = 1;
+    /// Makes a table of with the specified number of players.
+    pub fn new(number_of_players: usize, evaluator: Arc<Evaluator>) -> Self {
+        if number_of_players > 23 {
+            panic!("Too many players for one table!")
+        }
+        let mut players = Vec::new();
+        for i in 0..number_of_players {
+            players.push(Player::new(i as i8))
+        }
+        let initial_index = number_of_players - 1;
+        let mut table = Table {
+            players,
+            evaluator,
+            flop: None,
+            turn: None,
+            river: None,
+            dealer_button_index: initial_index,
+            ante: 1,
+            hand_number: 0,
+            current_player_index: initial_index,
+            table_state: PreFlop,
+            player_bets: vec![0; number_of_players],
+            ante_round_increase: number_of_players as i32 * 2,
+        };
+        table.deal();
+        table
+    }
+
+    /// Reset the table state to the starting round state
+    fn reset_state_for_new_round(&mut self) {
+        self.table_state = PreFlop
+    }
+
+    /// Translates the flop into a human readable string
+    pub fn get_flop_string(&self) -> String {
+        match self.flop {
+            None => { "None".to_string() }
+            Some(cards) => { format!("{} {} {}", cards[0], cards[1], cards[2]) }
+        }
+    }
+
+    /// Translates the flop into a human readable string
+    pub fn get_flop_string_secret(&self) -> String {
+        match self.flop {
+            None => { "None".to_string() }
+            Some(cards) => {
+                match &self.table_state {
+                    PreFlop => { "Hidden".to_string() }
+                    _ => { format!("{} {} {}", cards[0], cards[1], cards[2]) }
+                }
+            }
+        }
+    }
+
+
+    /// Translates the turn into a human readable string
+    pub fn get_turn_string(&self) -> String {
+        match self.turn {
+            None => { "None".to_string() }
+            Some(card) => { card.to_string() }
+        }
+    }
+
+    pub fn get_turn_string_secret(&self) -> String {
+        match self.turn {
+            None => { "None".to_string() }
+            Some(card) => {
+                match &self.table_state {
+                    PreFlop => { "Hidden".to_string() }
+                    Flop => { "Hidden".to_string() }
+                    _ => { card.to_string() }
+                }
+            }
+        }
+    }
+
+    /// Translates the river into a human readable string
+    pub fn get_river_string(&self) -> String {
+        match self.river {
+            None => { "None".to_string() }
+            Some(card) => { card.to_string() }
+        }
+    }
+
+    pub fn get_river_string_secret(&self) -> String {
+        match self.river {
+            None => { "None".to_string() }
+            Some(card) => {
+                match &self.table_state {
+                    River => { card.to_string() }
+                    _ => { "Hidden".to_string() }
+                }
+            }
+        }
+    }
+
+    /// Returns the number of players
+    pub fn get_player_count(&self) -> usize {
+        self.players.len()
+    }
+
+    /// Takes an action, could be recursive if the table needs no input
+    pub fn take_action(&mut self, hand_action: HandAction) {
+        // If the game is over print out a message, and do not take any actions
+        if self.is_game_over() {
+            println!("Game is over! Results are included below:\n{}", self.get_results());
+            return;
+        }
+        // Make sure the player is active, or increment and go to the next player
+        if let PlayerState::Active(active) = self.get_current_player().player_state {
+            self.take_provided_action(hand_action, active);
+        } else {
+            panic!("Tried to take an action on a dead player");
+        }
+        // If there is only 1 alive player evaluate the winner
+        if self.get_alive_player_count() == 1 {
+            self.resolve_hand();
+            return;
+        }
+        // The round of betting is over
+        if self.is_betting_over() {
+            // The showdown is occurring, pick the winner
+            if self.table_state == River {
+                self.resolve_hand();
+                return;
+            }
+            // Move to the next betting stage
+            self.table_state.next_stage();
+            // Reset the turn to the next person alive past the deal index
+            self.current_player_index = self.dealer_button_index;
+            self.update_current_player_index_to_next_active();
+            // set everyone to not have a turn yet
+            for player in &mut self.players {
+                player.has_had_turn_this_round = false;
+            }
+        }
+    }
+
+    fn take_provided_action(&mut self, hand_action: HandAction, active_state: ActiveState) {
+        let difference = self.get_largest_active_bet() - active_state.current_bet;
+
+        // Now check how to advance the hand
+        match hand_action {
+            HandAction::Fold => {
+                self.get_current_player().fold();
+            }
+            HandAction::Check => {
+                // All in already, so stay all in
+                if difference == 0 {
+                    self.get_current_player().bet(0);
+                } else {
+                    self.get_current_player().fold();
+                }
+            }
+            HandAction::Call => {
+                let bet_amount = self.get_current_player().bet(difference);
+                let index = self.get_current_player().get_id() as usize;
+                *self.player_bets.get_mut(index).unwrap() += bet_amount;
+            }
+            HandAction::Raise(raise_amount) => {
+                // Ensure the bet isn't larger than the pot limit (pot + amount required to call)
+                let acceptable_bet = min(raise_amount + difference, self.get_pot_size() + difference);
+                let bet_amount = self.get_current_player().bet(acceptable_bet);
+                let index = self.get_current_player().get_id() as usize;
+                *self.player_bets.get_mut(index).unwrap() += bet_amount;
+            }
+        }
+        // Update to point at the next player
+        self.update_current_player_index_to_next_active();
+    }
+
+    pub fn get_pot_size(&self) -> i32 {
+        self.player_bets.iter().sum::<i32>()
+    }
+
+    pub fn get_results(&self) -> String {
+        let mut players_copy = self.players.clone();
+        players_copy.sort_by(|a, b| b.cmp(a));
+        let mut rank = 1;
+        let mut result_string = format!("Results:\nRank: {}, Player: {}\n", rank, players_copy.get(0).unwrap());
+        for (i, player) in players_copy.iter().skip(1).enumerate() {
+            // The players didn't tie, so increase the rank
+            if player != players_copy.get(i).unwrap() {
+                rank = i + 2;
+            }
+            result_string += &format!("Rank: {rank}, Player: {player}\n");
+        }
+        result_string
+    }
+
+    /// Gets the current turn information
+    pub fn get_current_turn_information(&mut self) -> JsonValue {
+        self.get_state_string_for_player(self.players.get(self.current_player_index).unwrap().get_id())
+    }
+
+    /// Cleans all the cards from the table
+    pub fn clean_table(&mut self) {
+        self.flop = None;
+        self.turn = None;
+        self.river = None;
+        for player in &mut self.players {
+            player.fold();
+        }
+    }
+
+    pub fn is_game_over(&self) -> bool {
+        let alive_player_count = self.players.iter().map(|x| i8::from(x.is_alive())).reduce(|x, y| x + y).unwrap();
+        alive_player_count == 1
+    }
+
+    /// Deals cards to all players that are still alive,
+    /// marks any dead players with their death turn number,
+    /// moves the dealer chip,
+    /// update all 5 table cards,
+    pub fn deal(&mut self) {
+        // If the game is over do not do anything
+        if self.is_game_over() {
+            return;
+        }
+        // Increment the hand number
+        self.hand_number += 1;
+        // Reset all player bets to zero
+        self.player_bets = vec![0; self.players.len()];
+        // Reset the state for a new round of betting
+        self.reset_state_for_new_round();
+        // Check all players for death
+        self.check_for_player_death();
+        // Make a deck
+        let deck = Card::generate_shuffled_deck();
+        let mut deck_iterator = deck.iter();
+        // Deal cards to the players and the table
+        self.deal_table_cards(&mut deck_iterator);
+        self.deal_player_cards_collect_ante(&mut deck_iterator);
+        // Find the next alive player index for dealer button
+        self.find_next_deal_button_index_and_update_current_player();
+        // If it is time to increase the ante do so.
+        if (self.hand_number) % self.ante_round_increase == 0 {
+            self.ante += Table::ANTE_INCREASE_AMOUNT;
+        }
+    }
+
+    /// Finds the next dealer button index (next player in the list that is alive
+    fn find_next_deal_button_index_and_update_current_player(&mut self) {
+        for _ in 0..self.players.len() {
+            // set the next dealer index by finding the next alive player
+            self.dealer_button_index += 1;
+            if self.dealer_button_index + 1 >= self.get_player_count() {
+                self.dealer_button_index = 0;
+            }
+            if self.players.get(self.dealer_button_index).unwrap().is_alive() {
+                break;
+            }
+        }
+        // Set the current dealer button, and then increment that
+        self.current_player_index = self.dealer_button_index;
+        self.update_current_player_index_to_next_active();
+    }
+
+    fn update_current_player_index_to_next_active(&mut self) {
+        for _ in 0..self.players.len() {
+            // Set the next active player to the next index, looping if too large
+            self.current_player_index += 1;
+            if self.current_player_index >= self.players.len() {
+                self.current_player_index = 0;
+            }
+            // An all in player is no longer can take actions so skip them as well
+            if self.get_current_player().total_money == 0 {
+                continue;
+            }
+            // If the player is active while also not all in they are the next active player
+            match self.players.get(self.current_player_index).unwrap().player_state {
+                PlayerState::Folded => {}
+                PlayerState::Active(_) => { break; }
+            }
+        }
+    }
+
+    /// Deals cards for the flop, turn, and river
+    fn deal_table_cards(&mut self, deck_iterator: &mut Iter<Card>) {
+        self.flop = Option::from([
+            *deck_iterator.next().unwrap(),
+            *deck_iterator.next().unwrap(),
+            *deck_iterator.next().unwrap(),
+        ]);
+        self.turn = Option::from(*deck_iterator.next().unwrap());
+        self.river = Option::from(*deck_iterator.next().unwrap());
+    }
+
+    /// Mark all players that died from the last round as dead now
+    fn check_for_player_death(&mut self) {
+        // Check if players died on the past round
+        for player in &mut self.players {
+            if player.death_hand_number.is_none() && player.total_money < self.ante {
+                player.death_hand_number = Some(self.hand_number)
+            }
+        }
+    }
+
+    /// Deal cards to the alive players and collect the ante from them.
+    fn deal_player_cards_collect_ante(&mut self, deck_iterator: &mut Iter<Card>) {
+        // Deal every alive player cards now
+        for (i, player) in &mut self.players.iter_mut().enumerate() {
+            if player.is_alive() {
+                let card1 = *deck_iterator.next().unwrap();
+                let card2 = *deck_iterator.next().unwrap();
+                player.deal([card1, card2]);
+                *self.player_bets.get_mut(i).unwrap() += player.bet(self.ante);
+                // the ante doesn't count as a turn so clarify the bot hasn't had a turn
+                player.has_had_turn_this_round = false;
+            } else {
+                player.player_state = PlayerState::Folded;
+            }
+        }
+    }
+
+    pub fn get_state_string_for_player(&self, id: i8) -> JsonValue {
+        let player_strings: Vec<_> = self.players.iter().map(|x| if x.get_id() == id { x.as_json() } else { x.as_json_no_secret_data() }).collect();
+        object! {
+            flop: self.get_flop_string_secret(),
+            turn: self.get_turn_string_secret(),
+            river: self.get_river_string_secret(),
+            dealer_button_index: self.dealer_button_index,
+            players: player_strings,
+            hand_number: self.hand_number,
+        }
+    }
+    fn is_betting_over(&self) -> bool {
+        let all_players_ready = self.check_all_players_ready_for_next_round();
+        let all_players_equal_bet = self.check_all_active_players_same_bet();
+        all_players_ready && all_players_equal_bet
+    }
+    fn get_largest_active_bet(&self) -> i32 {
+        self.players.iter().map(|x| match x.player_state {
+            PlayerState::Folded => { 0 }
+            PlayerState::Active(active_state) => { active_state.current_bet }
+        }).max().unwrap()
+    }
+    fn check_all_players_ready_for_next_round(&self) -> bool {
+        self.players.iter().map(|x| match x.player_state {
+            PlayerState::Folded => { true }
+            PlayerState::Active(_) => { x.has_had_turn_this_round || x.total_money == 0 }
+        }).reduce(|x, y| x && y).unwrap()
+    }
+    fn check_all_active_players_same_bet(&self) -> bool {
+        let max_bet = self.get_largest_active_bet();
+        self.players.iter().map(|x| match x.player_state {
+            PlayerState::Folded => { true }
+            PlayerState::Active(a) => {
+                x.total_money == 0 || a.current_bet == max_bet
+            }
+        }
+        ).reduce(|x, y| x && y).unwrap()
+    }
+    fn get_alive_player_count(&self) -> usize {
+        self
+            .players
+            .iter()
+            .map(|x| match x.player_state {
+                PlayerState::Folded => { 0 }
+                PlayerState::Active(_) => { 1 }
+            })
+            .reduce(|x, y| x + y)
+            .unwrap()
+    }
+
+    ///
+    /// Returns the difference between each players bets,
+    /// checks they're sorted lowest to highest bet amount so the vector will be all positive
+    /// # Arguments
+    ///
+    /// * `players`: A list of players sorted by their bet amounts
+    ///
+    /// returns: Vec<i32> The difference between all the current bets
+    ///
+    fn get_bet_increases_amount(players: &[Player]) -> Vec<i32> {
+        // Check that the slice is sorted
+        let bets: Vec<i32> = players.iter().map(|x| match x.player_state {
+            PlayerState::Folded => { panic!("Passed a folded player.") }
+            PlayerState::Active(a) => { a.current_bet }
+        }).collect();
+        if bets.windows(2).any(|w| w[0] > w[1]) {
+            panic!("Players are not sorted by their bets.")
+        }
+        let mut return_vector = vec![0; players.len()];
+        let mut prev_bet = 0;
+        for (i, player) in players.iter().enumerate() {
+            if let PlayerState::Active(active) = player.player_state {
+                let push_back_amount = active.current_bet - prev_bet;
+                prev_bet = active.current_bet;
+                return_vector[i] = push_back_amount;
+            }
+        }
+        return_vector
+    }
+    /// Picks winner(s), gives out winnings, and deals a new hand
+    fn resolve_hand(&mut self) {
+        // This is the everyone but one person has folded case, give that person the winnings
+        if self.get_alive_player_count() == 1 {
+            self.players.iter_mut().find(|x| match x.player_state {
+                PlayerState::Folded => { false }
+                PlayerState::Active(_) => { true }
+            }).unwrap().total_money += self.get_pot_size();
+        } else {
+            // Otherwise we need to give out winnings based on hand strength
+            let sorted_players = self.get_hand_result();
+            for mut list_of_players in sorted_players {
+                // Sort by the smallest bet to the largest bet
+                list_of_players.sort_by(Table::compare_players_by_bet_amount);
+                // filter out any folded players just in case
+                let list_of_players: Vec<Player> = list_of_players.into_iter().filter(|x| (*x).player_state.is_active()).collect();
+                let mut player_size = list_of_players.len() as i32;
+                let bet_amounts = Table::get_bet_increases_amount(&list_of_players);
+                for (i, bet_amount) in bet_amounts.iter().enumerate() {
+                    if self.get_pot_size() == 0 {
+                        break;
+                    }
+                    // Take the bet from everyone
+                    let mut total = 0;
+                    for bet in &mut self.player_bets {
+                        let side_pot_amount = min(*bet_amount, *bet);
+                        *bet -= side_pot_amount;
+                        total += side_pot_amount;
+                    }
+                    let total = total;
+                    let each_player_payout = total / player_size;
+                    let remainder = total % player_size;
+                    for j in i..list_of_players.len() {
+                        let winning_id = list_of_players[j].get_id();
+                        let winner = self.players.iter_mut().find(|x| x.get_id() == winning_id).unwrap();
+                        winner.total_money += each_player_payout;
+                        if (j as i32) < remainder {
+                            winner.total_money += 1;
+                        }
+                    }
+                    player_size -= 1;
+                }
+            }
+        }
+        self.deal();
+    }
+
+    fn compare_players_by_bet_amount(player1: &Player, player2: &Player) -> Ordering {
+        let player_states = (player1.player_state, player2.player_state);
+        match player_states {
+            (PlayerState::Folded, PlayerState::Folded) => { player1.get_id().cmp(&player2.get_id()) }
+            (PlayerState::Active(_), PlayerState::Folded) => { Ordering::Greater }
+            (PlayerState::Folded, PlayerState::Active(_)) => { Ordering::Less }
+            (PlayerState::Active(one), PlayerState::Active(two)) => { one.current_bet.cmp(&two.current_bet) }
+        }
+    }
+
+    fn get_current_player(&mut self) -> &mut Player {
+        self.players.get_mut(self.current_player_index).unwrap()
+    }
+
+    pub fn compare_players(&self, shared_cards: &[Card], player1: &Player, player2: &Player) -> Ordering {
+        if !player1.player_state.is_active() && !player2.player_state.is_active() {
+            return Ordering::Equal;
+        } else if !player1.player_state.is_active() {
+            return Ordering::Greater;
+        } else if !player2.player_state.is_active() {
+            return Ordering::Less;
+        }
+
+        let mut a_set: Vec<Card> = shared_cards.into();
+        if let PlayerState::Active(a) = &player1.player_state {
+            a_set.extend(a.hand.iter());
+        }
+
+        let mut b_set: Vec<Card> = shared_cards.into();
+        if let PlayerState::Active(b) = &player2.player_state {
+            b_set.extend(b.hand.iter());
+        }
+
+        let a = self.evaluator.evaluate(a_set).expect("Couldn't evaluate hand 1");
+        let b = self.evaluator.evaluate(b_set).expect("Couldn't evaluate hand 2");
+        b.cmp(&a)
+    }
+
+    pub fn sort_by_hands(&self, total_hand: &[Card], alive_players: &mut [Player]) {
+        alive_players.sort_by(|player1, player2| {
+            self.compare_players(total_hand, player1, player2)
+        });
+    }
+
+    pub fn get_hand_result(&self) -> Vec<Vec<Player>> {
+        let mut players_copy = self.players.clone();
+
+        let total_hand = vec![*self.flop.unwrap().get(0).unwrap(), *self.flop.unwrap().get(1).unwrap(), *self.flop.unwrap().get(2).unwrap(), self.turn.unwrap(), self.river.unwrap()];
+        self.sort_by_hands(&total_hand, &mut players_copy);
+        let mut rankings = Vec::new();
+        rankings.push(Vec::new());
+        rankings[0].push(players_copy[0]);
+        for curr_player in players_copy.iter().skip(1) {
+            if self.compare_players(&total_hand, curr_player, &rankings[rankings.len() - 1][0]).is_gt() {
+                rankings.push(Vec::new());
+            }
+            let rankings_size = rankings.len();
+            rankings[rankings_size - 1].push(*curr_player);
+        }
+        rankings
+    }
+}
